refactor: parse produces a linked list of symbols
codegen now iterates through the symbols generated during parse and
codegens them. The intent behind this is to soon generate symbols for
string literals and codegen those. As a side-effect, scc is now capable
of defining and calling limited functions, though it does not type check
them and assumes they return int.