refactor: improve declaration parsing
this commit significantly beefs up and standardizes the parsing of
declarations. With this commit we can now parse declarations of
functions, which allows us to finally silence the warnings in our test
suite by forward-declaring the functions that we call.

Function prototypes are not type-checked yet.

This commit also introduces the notion of linkage. Function prototypes
have external linkage by default, which means they get no codegen.