bug: align the stack pointer when calling functions
System V requires that the stack pointer be aligned on a 16-byte
boundary upon function call. scc uses a dynamic value stack codegen
strategy and is constantly pushing and popping values from the stack; it
has no idea what the stack pointer looks like at the time of a call.

Because of this, we must manually align our stack pointer before each
call and restore it upon return.

This allows us to call printf without segfaulting, since printf performs
a read off the stack that will trap if unaligned.